Declare the projection constants unsigned in math.c
projectiondata1..10 are `dw` in dseg (asmorig/dseg.asm:20306-20315) and
shape3d.c declares all ten of them unsigned, which is also how set_projection
builds them:

    projectiondata9 = (long)cos_fast(projectiondata1) * projectiondata3
                    / sin_fast(projectiondata1);

math.c declared four of them as int instead, so the same 16-bit word was
signed in one translation unit and unsigned in the other. That mattered for
projectiondata9 and projectiondata10: the expression above is a cotangent
scaled by the half-width, so it grows without bound as the projection angle
narrows and passes 7FFFh well before anything else goes wrong. Past that
point math.c read them as negative and the projection multiply in
vector_to_point came out with the wrong sign, where the original's `mul` is
unsigned.

The use sites already cast around it, so this changes no behaviour on its
own - it removes the trap rather than the symptom.

projectiondata5 and projectiondata8 are a different case and the declaration
alone would have changed them. They are consumed by

        add     ax, projectiondata5
        jo      short loc_3243C

and jo reads both operands as signed words, so they have to reach
saturate_projection sign-extended. As unsigned they would have widened by
zero extension instead and a value of 8000h or more would have saturated the
wrong way. Cast them back to int at the four call sites, which also documents
which of the two roles each constant plays.

+++ src/restunts/c/math.c
@@ -747,7 +747,16 @@ int vector_op_unk2(struct VECTOR* vec) {
 	return result;
 }
 
-extern int projectiondata5, projectiondata8, projectiondata9, projectiondata10;
+// All ten of these are `dw` in dseg and are declared unsigned in shape3d.c,
+// which is also how set_projection produces them. math.c used to declare this
+// subset as int, so projectiondata9/10 - the only ones that can grow past
+// 7FFFh - reached the projection multiply as negative values.
+//
+// The two roles are not the same, and the casts at the use sites say which is
+// which: 9 and 10 are the operands of `mul`, so they stay unsigned there,
+// while 5 and 8 are added with `add ax, .. / jo`, whose overflow test reads
+// both operands as signed words.
+extern unsigned projectiondata5, projectiondata8, projectiondata9, projectiondata10;
 
 
 // Each `add ax, projectiondataN` in the original is followed by `jo`, and on
@@ -787,7 +796,7 @@ void vector_to_point(struct VECTOR* vec, struct POINT2D* outpt) {
 		comp = (short)(proj >> 15);
 
 		if (vec->z > comp) { 
-			outpt->px = saturate_projection(-(long)(proj / (unsigned short)vec->z) + projectiondata5);
+			outpt->px = saturate_projection(-(long)(proj / (unsigned short)vec->z) + (int)projectiondata5);
 		} else
 			outpt->px = -0x7D00;
 	} else {
@@ -795,7 +804,7 @@ void vector_to_point(struct VECTOR* vec, struct POINT2D* outpt) {
 		comp = (short)(proj >> 15);
 
 		if (vec->z > comp) 
-			outpt->px = saturate_projection((long)(proj / (unsigned short)vec->z) + projectiondata5);
+			outpt->px = saturate_projection((long)(proj / (unsigned short)vec->z) + (int)projectiondata5);
 		else
 			outpt->px = 0x7D00;
 	}
@@ -805,7 +814,7 @@ void vector_to_point(struct VECTOR* vec, struct POINT2D* outpt) {
 		comp = (short)(proj >> 15);
 
 		if (vec->z > comp) 
-			outpt->py = saturate_projection((long)(proj / (unsigned short)vec->z) + projectiondata8);
+			outpt->py = saturate_projection((long)(proj / (unsigned short)vec->z) + (int)projectiondata8);
 		else
 			outpt->py = 0x7D00;
 	} else {
@@ -813,7 +822,7 @@ void vector_to_point(struct VECTOR* vec, struct POINT2D* outpt) {
 		comp = (short)(proj >> 15);
 
 		if (vec->z > comp) 
-			outpt->py = saturate_projection(-(long)(proj / (unsigned short)vec->z) + projectiondata8);
+			outpt->py = saturate_projection(-(long)(proj / (unsigned short)vec->z) + (int)projectiondata8);
 		else
 			outpt->py = -0x7D00;
 	}
